% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performances.R
\name{perfPrecision}
\alias{perfPrecision}
\title{perfRepet}
\usage{
perfPrecision(df, x, run, day)
}
\arguments{
\item{df}{dataframe}

\item{x}{column name of the value to evaluate}

\item{run}{column name where run ids are stored}

\item{day}{column name where days ids are stored}
}
\value{
dataframe containing main statistics on variance
}
\description{
Computes Assay Precision according to CLSI EP05-A3.
}
\details{
references: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2556577/
}
\examples{
data(Glucose,package="VCA")
perfPrecision(Glucose, 'result','run', 'day')

}
