% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performances.R
\name{perfRepet}
\alias{perfRepet}
\title{perfRepet}
\usage{
perfRepet(df, x, run, day)
}
\arguments{
\item{df:}{dataframe}

\item{x:}{column name of value to evaluate}

\item{day:}{column name where nuber of days is stored}
}
\value{
dataframe containing main statistics on variance
}
\description{
Computes Assay Precision according to CLSI EP05-A3.
}
\details{
references: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2556577/
}
\examples{
data(Glucose,package="VCA")
perfRepet(Glucose, 'result','run', 'day')

}
