% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPrecision}
\alias{plotPrecision}
\title{Plot Precision Chart}
\usage{
plotPrecision(df, x, run, day)
}
\arguments{
\item{df}{a datatable}

\item{x}{string, column name of the results}

\item{run}{string, column name of the runs ID}

\item{day}{string, column name of the Days ID}
}
\value{
ggplot object
}
\description{
A plotting format for plotting assay's precision
}
\details{
Middle line is the grand mean, short lines are daily's mean, red points are run means and black point individual observations.
}
\examples{
data(Glucose,package="VCA")
plotPrecision(Glucose, 'result','run', 'day')
}
