% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performances.R
\name{perfPrecision}
\alias{perfPrecision}
\title{perfPrecision}
\usage{
perfPrecision(df, x, run, day, by = NULL, site = NULL)
}
\arguments{
\item{df}{dataframe}

\item{x}{string. column name of the value to evaluate}

\item{run}{string. column name where run ids are stored}

\item{day}{string. column name where days ids are stored}

\item{by}{string, default to NULL. column name of a slicing variable. Use by if the table contains a variable to repeat the analysis on (for example a QC level or a gold-standard identification variable). If by is used, a list containing one result per slice level is returned.}

\item{site}{string, column name where sites ids are stored. If site is not null, the outputs gives back reproducibility performances.}
}
\value{
dataframe containing main statistics on variance
}
\description{
Computes Assay Precision according to CLSI EP05-A3.
}
\details{
Heavily relies on the VCA package for computing the variance component analysis.
references: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2556577/
}
\examples{
data(Glucose,package="VCA")
perfPrecision(Glucose, 'result','run', 'day')

}
